requirements.txt
# InfographAI Phase 1 Dependencies

# PowerPoint generation
python-pptx>=0.6.21

# Text measurement
Pillow>=10.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# LLM Integration
anthropic>=0.40.0

# Color extraction (for brand engine)
scikit-learn>=1.3.0
numpy>=1.24.0

# Future dependencies (Phase 2+)
# litellm>=1.0.0
